feat(parser)!: TS 抽出層を oxc-parser に切り替え、ts-morph を完全削除 (closes #159)
フェーズ2(切り替え)。typescript-oxc.ts の oxc バックエンドを
src/parsers/typescript.ts 本体に統合し、公開 API(createTSParser /
parseTSFilePaths / globCodeFiles / DEFAULT_ID_TOKEN ほか)はそのまま、
実装だけを差し替える。出力契約(nodes/edges の配列順序・contentHash)は
ts-morph 版とビット単位で同一なので既存 .trace.lock は不変(INV-L4)。

- ts-morph を dependencies から削除(devDependency にも残さない)
- 差分テストは ts-morph オラクル喪失に伴い、フェーズ1で検証済みの挙動を
  期待値として明示的にピン留めする回帰テストへ転換
  (tests/typescript-oxc-regression.test.ts、R1〜R11 の11観点:
  拡張子優先順と .tsx シャドーイング / .mjs・.cjs 置換 / package.json
  types・typings・main / script ターゲット抑止 / jsx・allowJs・
  resolveJsonModule のみが解決に影響(JSONC+extends 込み) / binder の
  functions-first 順 / 宣言種別ごとの getText() スパン / DirectoryCache
  BFS ファイル順 / BOM 除去 / export {} from / 構文エラー時の module
  record フォールバックと symbol ノード非復元)
- builder.ts の tsEnvKey はそのまま維持: tsconfig 内容ハッシュは
  「ts-morph が読むから」ではなく「新リゾルバが jsx / allowJs /
  resolveJsonModule を tsconfig から読むから」必要 — コメントを実態に更新
- cli.ts の lazy-import 説明コメントも oxc-parser 実態に更新

計測(このリポジトリ、ARTGRAPH_CACHE=0、warmup 1 + 9回の中央値):
- コールド check 全体: 2907ms(main/ts-morph)→ 1512ms
- TS 抽出層単体(parseTSFilePaths, 118 files, in-process 中央値):
  956ms → 184ms(+ モジュールロード ~300ms → ~15ms)
- ユニットテストスイート全体: 71.6s → 13.0s
