+++ package.json
@@ -63,7 +63,6 @@
     "mdast-util-to-string": "^4.0.0",
     "oxc-parser": "^0.137.0",
     "remark-parse": "^11.0.0",
-    "ts-morph": "^25.0.0",
     "unified": "^11.0.5",
     "unist-util-visit": "^5.0.0",
     "yaml": "^2.9.0"
