+++ src/parsers/typescript.ts
@@ -1,21 +1,56 @@
-import type { Project, SourceFile } from "ts-morph";
-import { createHash } from "node:crypto";
 import { createRequire } from "node:module";
-import { existsSync } from "node:fs";
-import { resolve, relative } from "node:path";
+import { createHash } from "node:crypto";
+import { readFileSync, statSync } from "node:fs";
+import { dirname, resolve, relative } from "node:path";
 import fastGlob from "fast-glob";
 import type { GraphNode, GraphEdge } from "../types.js";
 import { NAMESPACED_ID_TOKEN } from "../req-id.js";
 
-// ts-morph is a CJS package and by far the heaviest import in the CLI
-// (~300 ms module load). Loading it lazily via createRequire keeps this
-// module's import cheap and — combined with the parse cache — lets a
-// fully-warm scan skip ts-morph entirely. `require` of a CJS dep is
-// synchronous, so callers (buildGraph is sync) don't need to change shape.
+// oxc-parser based TypeScript extraction layer (issue #159).
+//
+// This module replaced its original ts-morph backend with oxc's native
+// parser. The output contract is bit-for-bit what the ts-morph backend
+// produced — node/edge ARRAY ORDER and contentHash values included — so
+// existing `.trace.lock` files stay byte-identical across the swap (INV-L4).
+// That behavior was established by differential-testing both backends over
+// this repository, the test fixtures, and edge-case probes, and is now
+// pinned by tests/typescript-oxc-regression.test.ts. Three pieces of
+// compiler behavior are re-derived here from those empirical probes:
+//
+//   1. Relative import/export specifier resolution (the checker's behavior
+//      behind `getModuleSpecifierSourceFile`). The outcome depends only on
+//      the tsconfig's jsx / allowJs / resolveJsonModule — NOT on
+//      moduleResolution or package.json "type" (node16 strictness is not
+//      enforced on that API), and the resolved target must itself be a
+//      module (a script target yields no edge). See resolveRelativeImport.
+//   2. Symbol-mode export enumeration (`getExportedDeclarations`). Its Map
+//      iteration order is the checker's export-symbol-table order, which the
+//      binder fills FUNCTIONS FIRST (top-level function statements in source
+//      order, then everything else in source order — bindEachFunctionsFirst),
+//      and its getText() spans differ per declaration kind (statement span
+//      including `export` for functions/classes/…, declarator span excluding
+//      `export const` for variables). See extractSymbols.
+//   3. Import-clause enumeration for symbol-mode edges. Walked from the AST;
+//      for files with fatal syntax errors (where oxc returns an empty
+//      program) the parser's error-tolerant module record still carries the
+//      import/export shape and is used as a fallback. Known limit: symbol
+//      NODES cannot be recovered from such files (the TS compiler recovered
+//      partial declarations; oxc does not).
+//
+// oxc-parser is a CJS package with a native binding; loading it lazily via
+// createRequire keeps this module's import cheap — combined with the parse
+// cache, a fully-warm scan never loads the parser at all. `require` of a CJS
+// dep is synchronous, so callers (buildGraph is sync) don't need to change
+// shape.
 const requireCjs = createRequire(import.meta.url);
-function loadTsMorph(): typeof import("ts-morph") {
-  return requireCjs("ts-morph") as typeof import("ts-morph");
+let oxcModule: typeof import("oxc-parser") | undefined;
+function loadOxc(): typeof import("oxc-parser") {
+  return (oxcModule ??= requireCjs("oxc-parser") as typeof import("oxc-parser"));
 }
+
+type OxcParseResult = import("oxc-parser").ParseResult;
+type OxcProgram = OxcParseResult["program"];
+type OxcStatement = OxcProgram["body"][number];
 
 // Default requirement-ID *token* used when no custom `reqPatterns.codeId` is set.
 // The token matches the whole ID (e.g. `FR-001`, `auth/AUTH-2`, `Requirement-3`).
@@ -28,9 +63,6 @@ export const DEFAULT_ID_TOKEN = NAMESPACED_ID_TOKEN;
 // Regexes that locate requirement IDs in code/test tags. When the project sets a
 // custom `reqPatterns.codeId`, these are rebuilt from that token so that @impl /
 // test-bracket / `req:` annotations track the same IDs the markdown parser emits.
-// Exported (with buildIdMatchers/extractImplTags/hash below) so the oxc-based
-// parser in typescript-oxc.ts shares the exact same tag grammar and hashing —
-// the regex layer is AST-independent and must not fork between the two backends.
 export interface IdMatchers {
   implRe: RegExp;
   reqIdRe: RegExp;
@@ -61,35 +93,32 @@ export interface SymbolRange {
   endLine: number;
 }
 
-function buildProjectOptions(rootDir: string) {
-  const tsconfigPath = resolve(rootDir, "tsconfig.json");
-  return existsSync(tsconfigPath)
-    ? { tsConfigFilePath: tsconfigPath, skipAddingFilesFromTsConfig: true }
-    : { skipAddingFilesFromTsConfig: true };
-}
-
 export function createTSParser(
   rootDir: string,
   patterns: string[],
   mode: "file" | "symbol" = "file",
   codeId?: string,
 ) {
-  const { Project: TsProject } = loadTsMorph();
-  const project = new TsProject(buildProjectOptions(rootDir));
-
-  for (const pattern of patterns) {
-    project.addSourceFilesAtPaths(resolve(rootDir, pattern));
-  }
-
   const matchers = buildIdMatchers(codeId);
-  return { project, parse: () => parseProject(project, rootDir, mode, matchers) };
+  return {
+    parse: (): ParsedTS => {
+      const ctx = createResolverContext(rootDir);
+      const nodes: GraphNode[] = [];
+      const edges: GraphEdge[] = [];
+      for (const filePath of enumerateFiles(rootDir, patterns)) {
+        const parsed = parseTSFile(filePath, rootDir, mode, matchers, ctx);
+        nodes.push(...parsed.nodes);
+        edges.push(...parsed.edges);
+      }
+      return { nodes, edges };
+    },
+  };
 }
 
-// Resolve the code-file set for `patterns` with the exact glob call ts-morph's
-// RealFileSystemHost makes inside `addSourceFilesAtPaths` (fast-glob, cwd =
+// Resolve the code-file set for `patterns` in one fast-glob call (cwd =
 // process.cwd(), absolute results). The parse-cache path discovers the file
 // set through this helper so warm runs see byte-for-byte the same set a full
-// ts-morph scan would, without loading ts-morph.
+// createTSParser scan enumerates, without loading the parser.
 export function globCodeFiles(rootDir: string, patterns: string[]): string[] {
   return fastGlob.sync(
     patterns.map((p) => resolve(rootDir, p).replace(/\\/g, "/")),
@@ -98,60 +127,178 @@ export function globCodeFiles(rootDir: string, patterns: string[]): string[] {
 }
 
 // Parse exactly the given files (used by the parse-cache path to reparse only
-// changed files). Files are all added to one Project before parsing, mirroring
-// createTSParser's add-all-then-parse flow; import resolution consults the
-// real file system, so targets outside `filePaths` still resolve the same way
-// they do in a full scan. Returns a fragment per input path.
+// changed files). Import resolution consults the real file system, so targets
+// outside `filePaths` still resolve the same way they do in a full scan.
+// Returns a fragment per input path.
 export function parseTSFilePaths(
   rootDir: string,
   filePaths: string[],
   mode: "file" | "symbol" = "file",
   codeId?: string,
 ): Map<string, ParsedTS> {
-  const { Project: TsProject } = loadTsMorph();
-  const project = new TsProject(buildProjectOptions(rootDir));
   const matchers = buildIdMatchers(codeId);
-
-  const sourceFiles = filePaths.map((p) => project.addSourceFileAtPath(p));
+  const ctx = createResolverContext(rootDir);
   const out = new Map<string, ParsedTS>();
-  for (let i = 0; i < filePaths.length; i++) {
-    out.set(filePaths[i], parseTSSourceFile(sourceFiles[i], rootDir, mode, matchers));
+  for (const filePath of filePaths) {
+    out.set(filePath, parseTSFile(filePath, rootDir, mode, matchers, ctx));
   }
   return out;
 }
 
-function parseProject(
-  project: Project,
-  rootDir: string,
-  mode: "file" | "symbol",
-  matchers: IdMatchers,
-): ParsedTS {
-  const nodes: GraphNode[] = [];
-  const edges: GraphEdge[] = [];
-
-  for (const sourceFile of project.getSourceFiles()) {
-    const parsed = parseTSSourceFile(sourceFile, rootDir, mode, matchers);
-    nodes.push(...parsed.nodes);
-    edges.push(...parsed.edges);
+// One glob per pattern, deduped, then ordered the way the previous ts-morph
+// backend's project.getSourceFiles() iterated. Preserving that order keeps
+// node/edge output — and therefore `.trace.lock` bytes — stable across the
+// backend swap.
+function enumerateFiles(rootDir: string, patterns: string[]): string[] {
+  const seen = new Set<string>();
+  const files: string[] = [];
+  for (const pattern of patterns) {
+    const matches = fastGlob.sync(resolve(rootDir, pattern).replace(/\\/g, "/"), {
+      cwd: resolve(),
+      absolute: true,
+    });
+    for (const filePath of matches) {
+      if (!seen.has(filePath)) {
+        seen.add(filePath);
+        files.push(filePath);
+      }
+    }
   }
-
-  return { nodes, edges };
+  return orderByDirectoryDepth(files);
 }
 
-function parseTSSourceFile(
-  sourceFile: SourceFile,
+// Sibling directory / file comparator (ts-morph's LocaleStringComparer).
+function compareBaseNames(a: string, b: string): number {
+  const result = a.localeCompare(b, "en-us-u-kf-upper");
+  return result < 0 ? -1 : result === 0 ? 0 : 1;
+}
+
+function baseName(p: string): string {
+  return p.slice(p.lastIndexOf("/") + 1);
+}
+
+function parentDir(p: string): string {
+  const idx = p.lastIndexOf("/");
+  return idx <= 0 ? "/" : p.slice(0, idx);
+}
+
+// Replicate the legacy project.getSourceFiles() iteration order: the
+// DirectoryCache held the directories of added files (plus the intermediate
+// directories that connect a directory to an already-cached ancestor) and
+// walked them breadth-first by path depth — "orphan" root directories in
+// insertion order, child directories and files sorted by base name. Files
+// are added per pattern in glob order, which fixes the orphan insertion
+// order.
+function orderByDirectoryDepth(files: string[]): string[] {
+  const filesByDir = new Map<string, string[]>();
+  const cached = new Set<string>();
+  const orphans: string[] = [];
+
+  const removeOrphan = (path: string): void => {
+    const idx = orphans.indexOf(path);
+    if (idx !== -1) orphans.splice(idx, 1);
+  };
+
+  // DirectoryCache#addDirectory: adopt orphans that are direct children,
+  // become an orphan when the parent isn't cached, then connect any orphan
+  // descendants by filling the intermediate directories.
+  const addDirectory = (path: string): void => {
+    for (const orphan of orphans.filter((o) => o !== path && parentDir(o) === path)) {
+      removeOrphan(orphan);
+    }
+    const parent = parentDir(path);
+    if (parent !== path && !cached.has(parent) && !orphans.includes(path)) {
+      orphans.push(path);
+    }
+    cached.add(path);
+    for (const orphan of orphans.filter((o) => o !== path && o.startsWith(`${path}/`))) {
+      fillParents(orphan);
+    }
+  };
+
+  // DirectoryCache#fillParentsOfDirPath: create the intermediate directories
+  // between `dirPath` and its nearest cached ancestor (top-down) — nothing
+  // when no ancestor is cached.
+  const fillParents = (dirPath: string): void => {
+    const passed: string[] = [];
+    let current = dirPath;
+    let parent = parentDir(current);
+    while (current !== parent) {
+      current = parent;
+      parent = parentDir(current);
+      if (cached.has(current)) {
+        for (const p of passed) addDirectory(p);
+        break;
+      }
+      passed.unshift(current);
+    }
+  };
+
+  for (const filePath of files) {
+    const dir = parentDir(filePath);
+    if (!cached.has(dir)) {
+      fillParents(dir);
+      addDirectory(dir);
+    }
+    const dirFiles = filesByDir.get(dir);
+    if (dirFiles) dirFiles.push(filePath);
+    else filesByDir.set(dir, [filePath]);
+  }
+
+  const childDirs = new Map<string, string[]>();
+  for (const dir of cached) {
+    const parent = parentDir(dir);
+    if (parent === dir || !cached.has(parent)) continue;
+    const siblings = childDirs.get(parent);
+    if (siblings) siblings.push(dir);
+    else childDirs.set(parent, [dir]);
+  }
+  for (const siblings of childDirs.values()) {
+    siblings.sort((a, b) => compareBaseNames(baseName(a), baseName(b)));
+  }
+
+  // getSourceFilesByDirectoryDepth: BFS over depth levels, seeded with the
+  // orphans (insertion order), children appended as their parent is visited.
+  const depthOf = (p: string): number => p.split("/").length;
+  const levels = new Map<number, string[]>();
+  const enqueue = (dir: string): void => {
+    const depth = depthOf(dir);
+    const level = levels.get(depth);
+    if (level) level.push(dir);
+    else levels.set(depth, [dir]);
+  };
+  for (const orphan of orphans) enqueue(orphan);
+
+  const out: string[] = [];
+  if (levels.size === 0) return out;
+  let depth = Math.min(...levels.keys());
+  while (levels.size > 0) {
+    for (const dir of levels.get(depth) ?? []) {
+      const dirFiles = (filesByDir.get(dir) ?? []).sort((a, b) =>
+        compareBaseNames(baseName(a), baseName(b)),
+      );
+      out.push(...dirFiles);
+      for (const child of childDirs.get(dir) ?? []) enqueue(child);
+    }
+    levels.delete(depth);
+    depth++;
+  }
+  return out;
+}
+
+function parseTSFile(
+  filePath: string,
   rootDir: string,
   mode: "file" | "symbol",
   matchers: IdMatchers,
+  ctx: ResolverContext,
 ): ParsedTS {
   const nodes: GraphNode[] = [];
   const edges: GraphEdge[] = [];
 
-  const filePath = sourceFile.getFilePath();
   const relPath = relative(rootDir, filePath);
-
-  const fileContent = sourceFile.getFullText();
-  const fileHash = hash(fileContent);
+  const content = stripBom(readFileSync(filePath, "utf-8"));
+  const fileHash = hash(content);
 
   const isTest = /\.(test|spec)\.(ts|tsx)$/.test(filePath);
 
@@ -162,81 +309,436 @@ function parseTSSourceFile(
     contentHash: fileHash,
   });
 
-  let symbolRanges: SymbolRange[] = [];
-
-  if (mode === "symbol" && !isTest) {
-    symbolRanges = extractSymbols(sourceFile, relPath, nodes);
+  // parseSync never throws on syntax errors (it reports res.errors and may
+  // return an empty program); the guard covers pathological inputs only.
+  let parsed: OxcParseResult | undefined;
+  try {
+    parsed = loadOxc().parseSync(filePath, content);
+  } catch {
+    parsed = undefined;
   }
 
-  extractImports(sourceFile, relPath, rootDir, edges, mode, isTest);
-  extractImplTags(fileContent, relPath, isTest, edges, mode, symbolRanges, matchers);
+  let symbolRanges: SymbolRange[] = [];
+  if (parsed && mode === "symbol" && !isTest) {
+    symbolRanges = extractSymbols(parsed.program, content, relPath, nodes);
+  }
+  if (parsed) {
+    extractImports(parsed, content, relPath, filePath, rootDir, edges, mode, isTest, ctx);
+  }
+  extractImplTags(content, relPath, isTest, edges, mode, symbolRanges, matchers);
 
   return { nodes, edges };
 }
 
+// The TS compiler host strips a UTF-8 BOM when reading files — the file-hash
+// input and all node spans are relative to the stripped text.
+function stripBom(content: string): string {
+  return content.charCodeAt(0) === 0xfeff ? content.slice(1) : content;
+}
+
+// ---------------------------------------------------------------------------
+// Symbol-mode export extraction (getExportedDeclarations semantics)
+// ---------------------------------------------------------------------------
+
+// A top-level declaration a local name can resolve to. `start`/`end` delimit
+// the text ts-morph's getText() returned for the declaration node:
+//   - functions/classes/interfaces/enums/namespaces/type aliases: the whole
+//     statement, INCLUDING `export`/`default` when directly exported and any
+//     decorators (which may sit before the `export` keyword);
+//   - variables: the VariableDeclarator (`x = 1`) or the leaf BindingElement
+//     of a destructuring pattern — never `export const`.
+interface LocalDecl {
+  isFunction: boolean;
+  start: number;
+  end: number;
+}
+
+interface ExportEntry {
+  name: string;
+  start: number;
+  end: number;
+}
+
+function isFunctionDecl(node: { type: string } | null | undefined): boolean {
+  return node?.type === "FunctionDeclaration" || node?.type === "TSDeclareFunction";
+}
+
+// Statement-text start: the export wrapper's start, widened to include class
+// decorators (oxc keeps decorators that precede `export` outside the
+// statement span; TS's getText() includes them).
+function declTextStart(stmtStart: number, decl: unknown): number {
+  let start = stmtStart;
+  const decorators = (decl as { decorators?: Array<{ start: number }> }).decorators;
+  if (decorators) {
+    for (const d of decorators) {
+      if (d.start < start) start = d.start;
+    }
+  }
+  return start;
+}
+
 function extractSymbols(
-  sourceFile: SourceFile,
+  program: OxcProgram,
+  content: string,
   relPath: string,
   nodes: GraphNode[],
 ): SymbolRange[] {
-  const ranges: SymbolRange[] = [];
-  const exported = sourceFile.getExportedDeclarations();
+  const declMap = collectTopLevelDecls(program);
+  const lookup = (name: string): LocalDecl | undefined => {
+    const decls = declMap.get(name);
+    if (!decls || decls.length === 0) return undefined;
+    // Merged symbols (fn+namespace, overloads): the binder registers function
+    // declarations first, so `symbol.declarations[0]` is the first function
+    // when one exists, else the first declaration in source order.
+    return decls.find((d) => d.isFunction) ?? decls[0];
+  };
+
+  const entries: ExportEntry[] = [];
   const seen = new Set<string>();
+  const push = (name: string, start: number, end: number) => {
+    if (seen.has(name)) return;
+    seen.add(name);
+    entries.push({ name, start, end });
+  };
 
-  for (const [name, declarations] of exported) {
-    const symbolName = name === "default" ? "default" : name;
-    if (seen.has(symbolName)) continue;
-
-    const localDecl = declarations.find((d) => d.getSourceFile() === sourceFile);
-    if (!localDecl) continue;
-
-    seen.add(symbolName);
-    const symbolId = `symbol:${relPath}#${symbolName}`;
-    const symbolHash = hash(localDecl.getText());
-
-    nodes.push({
-      id: symbolId,
-      kind: "symbol",
-      filePath: relPath,
-      contentHash: symbolHash,
-    });
-
-    ranges.push({
-      name: symbolName,
-      startLine: localDecl.getStartLineNumber(),
-      endLine: localDecl.getEndLineNumber(),
-    });
+  // Pass 1 — top-level function statements, source order. The TS binder binds
+  // FunctionDeclaration statements before everything else
+  // (bindEachFunctionsFirst), which is what orders the checker's export
+  // symbol table and therefore getExportedDeclarations() iteration.
+  for (const stmt of program.body) {
+    if (stmt.type === "ExportNamedDeclaration" && isFunctionDecl(stmt.declaration)) {
+      const decl = stmt.declaration as { id: { name: string } | null; end: number };
+      if (decl.id) push(decl.id.name, stmt.start, decl.end);
+    } else if (stmt.type === "ExportDefaultDeclaration" && isFunctionDecl(stmt.declaration)) {
+      push("default", stmt.start, stmt.declaration.end);
+    }
   }
 
+  // Pass 2 — every other export form, source order.
+  for (const stmt of program.body) {
+    if (stmt.type === "ExportNamedDeclaration") {
+      if (stmt.source) continue; // re-exports resolve to foreign declarations
+      const decl = stmt.declaration;
+      if (decl) {
+        if (isFunctionDecl(decl)) continue; // pass 1
+        if (decl.type === "VariableDeclaration") {
+          for (const declarator of decl.declarations) {
+            collectBindingNames(declarator.id, declarator.start, declarator.end, push);
+          }
+        } else if (decl.type === "ClassDeclaration") {
+          if (decl.id) push(decl.id.name, declTextStart(stmt.start, decl), decl.end);
+        } else if (
+          decl.type === "TSInterfaceDeclaration" ||
+          decl.type === "TSTypeAliasDeclaration" ||
+          decl.type === "TSEnumDeclaration"
+        ) {
+          push(decl.id.name, stmt.start, decl.end);
+        } else if (decl.type === "TSModuleDeclaration" && decl.id.type === "Identifier") {
+          push(decl.id.name, stmt.start, decl.end);
+        }
+      } else {
+        // `export { a as b }` — the exported name is the alias; the
+        // declaration is the local target. Imported / undeclared targets are
+        // skipped (foreign declarations never became symbol nodes).
+        for (const spec of stmt.specifiers) {
+          const localName = moduleExportName(spec.local);
+          const target = localName === undefined ? undefined : lookup(localName);
+          if (!target) continue;
+          const exportedName = moduleExportName(spec.exported);
+          if (exportedName !== undefined) push(exportedName, target.start, target.end);
+        }
+      }
+    } else if (stmt.type === "ExportDefaultDeclaration") {
+      const decl = stmt.declaration;
+      if (isFunctionDecl(decl)) continue; // pass 1
+      if (decl.type === "ClassDeclaration") {
+        push("default", declTextStart(stmt.start, decl), decl.end);
+      } else if (decl.type === "TSInterfaceDeclaration") {
+        push("default", stmt.start, decl.end);
+      } else if (decl.type === "Identifier") {
+        // `export default foo;` resolves to foo's local declaration (skipped
+        // when foo is imported/undeclared).
+        const target = lookup(decl.name);
+        if (target) push("default", target.start, target.end);
+      } else {
+        // `export default <expr>;` — the symbol hashes the expression node.
+        push("default", decl.start, decl.end);
+      }
+    }
+    // ExportAllDeclaration / TSExportAssignment (`export =`): no local symbols.
+  }
+
+  const lineStarts = buildLineStarts(content);
+  const ranges: SymbolRange[] = [];
+  for (const entry of entries) {
+    nodes.push({
+      id: `symbol:${relPath}#${entry.name}`,
+      kind: "symbol",
+      filePath: relPath,
+      contentHash: hash(content.slice(entry.start, entry.end)),
+    });
+    ranges.push({
+      name: entry.name,
+      startLine: lineOf(lineStarts, entry.start),
+      endLine: lineOf(lineStarts, entry.end),
+    });
+  }
   return ranges;
 }
 
+function collectTopLevelDecls(program: OxcProgram): Map<string, LocalDecl[]> {
+  const map = new Map<string, LocalDecl[]>();
+  const add = (name: string, decl: LocalDecl) => {
+    const list = map.get(name);
+    if (list) list.push(decl);
+    else map.set(name, [decl]);
+  };
+
+  for (const stmt of program.body) {
+    // The declaration text of an export-wrapped node starts at the statement
+    // (`export …`), so unwrap but keep the statement start.
+    let decl: OxcStatement | NonNullable<unknown> = stmt;
+    const stmtStart = stmt.start;
+    if (stmt.type === "ExportNamedDeclaration" && stmt.declaration) {
+      decl = stmt.declaration;
+    } else if (
+      stmt.type === "ExportDefaultDeclaration" &&
+      (isFunctionDecl(stmt.declaration) ||
+        stmt.declaration.type === "ClassDeclaration" ||
+        stmt.declaration.type === "TSInterfaceDeclaration")
+    ) {
+      decl = stmt.declaration;
+    }
+
+    const node = decl as {
+      type: string;
+      start: number;
+      end: number;
+      id?: { type: string; name?: string } | null;
+      declarations?: Array<{ id: unknown; start: number; end: number }>;
+    };
+    switch (node.type) {
+      case "FunctionDeclaration":
+      case "TSDeclareFunction":
+        if (node.id?.name) {
+          add(node.id.name, { isFunction: true, start: stmtStart, end: node.end });
+        }
+        break;
+      case "ClassDeclaration":
+        if (node.id?.name) {
+          add(node.id.name, {
+            isFunction: false,
+            start: declTextStart(stmtStart, node),
+            end: node.end,
+          });
+        }
+        break;
+      case "TSInterfaceDeclaration":
+      case "TSTypeAliasDeclaration":
+      case "TSEnumDeclaration":
+        if (node.id?.name) {
+          add(node.id.name, { isFunction: false, start: stmtStart, end: node.end });
+        }
+        break;
+      case "TSModuleDeclaration":
+        if (node.id?.type === "Identifier" && node.id.name) {
+          add(node.id.name, { isFunction: false, start: stmtStart, end: node.end });
+        }
+        break;
+      case "VariableDeclaration":
+        for (const declarator of node.declarations ?? []) {
+          collectBindingNames(declarator.id, declarator.start, declarator.end, (name, s, e) =>
+            add(name, { isFunction: false, start: s, end: e }),
+          );
+        }
+        break;
+    }
+  }
+  return map;
+}
+
+// Walk a declarator's binding target down to leaf binding elements, mirroring
+// which node the checker reports as each name's declaration:
+//   - plain `x = 1`: the whole declarator;
+//   - object/array patterns: the leaf BindingElement (`a`, `b: c`, `d = 1`,
+//     `...rest`), recursing through nested patterns.
+function collectBindingNames(
+  id: unknown,
+  declaratorStart: number,
+  declaratorEnd: number,
+  add: (name: string, start: number, end: number) => void,
+): void {
+  const node = id as { type: string; name?: string; start: number; end: number };
+  if (node.type === "Identifier") {
+    if (node.name) add(node.name, declaratorStart, declaratorEnd);
+    return;
+  }
+  collectPatternNames(node, add);
+}
+
+function collectPatternNames(
+  pattern: unknown,
+  add: (name: string, start: number, end: number) => void,
+): void {
+  const node = pattern as {
+    type: string;
+    start: number;
+    end: number;
+    properties?: unknown[];
+    elements?: unknown[];
+  };
+  if (node.type === "ObjectPattern") {
+    for (const prop of node.properties ?? []) {
+      collectPatternMember(prop, add);
+    }
+  } else if (node.type === "ArrayPattern") {
+    for (const element of node.elements ?? []) {
+      if (element) collectPatternMember(element, add);
+    }
+  }
+}
+
+// One pattern member: a Property (object), a positional element (array), or a
+// RestElement. The member's own span is the BindingElement text that gets
+// hashed ("a", "b: c", "d = 1", "...rest"); nested patterns recurse to their
+// leaves instead.
+function collectPatternMember(
+  member: unknown,
+  add: (name: string, start: number, end: number) => void,
+): void {
+  const node = member as {
+    type: string;
+    start: number;
+    end: number;
+    name?: string;
+    value?: unknown;
+    argument?: unknown;
+    left?: unknown;
+  };
+  const leaf = (target: unknown) => {
+    const t = target as { type: string; name?: string; left?: unknown };
+    if (t.type === "Identifier") {
+      if (t.name) add(t.name, node.start, node.end);
+    } else if (t.type === "AssignmentPattern") {
+      const left = t.left as { type: string; name?: string };
+      if (left.type === "Identifier") {
+        if (left.name) add(left.name, node.start, node.end);
+      } else {
+        collectPatternNames(left, add);
+      }
+    } else {
+      collectPatternNames(t, add);
+    }
+  };
+
+  if (node.type === "Property") {
+    leaf(node.value);
+  } else if (node.type === "RestElement") {
+    leaf(node.argument);
+  } else {
+    leaf(node);
+  }
+}
+
+// `export { a as b }` names can be identifiers or string literals
+// (`export { a as "b-c" }`); the symbol name is the literal's VALUE.
+function moduleExportName(node: unknown): string | undefined {
+  const n = node as { type: string; name?: string; value?: unknown };
+  if (n.type === "Identifier") return n.name;
+  return typeof n.value === "string" ? n.value : undefined;
+}
+
+// ---------------------------------------------------------------------------
+// Import / re-export edges
+// ---------------------------------------------------------------------------
+
+interface ImportRef {
+  specifier: string;
+  hasDefault: boolean;
+  hasNamespace: boolean;
+  // Original (pre-alias) names of named imports, as source text —
+  // string-literal import names keep their quotes.
+  named: string[];
+}
+
 function extractImports(
-  sourceFile: SourceFile,
+  parsed: OxcParseResult,
+  content: string,
   relPath: string,
+  filePath: string,
   rootDir: string,
   edges: GraphEdge[],
-  mode: "file" | "symbol" = "file",
-  isTest: boolean = false,
-) {
+  mode: "file" | "symbol",
+  isTest: boolean,
+  ctx: ResolverContext,
+): void {
+  const importRefs: ImportRef[] = [];
+  const reexportSpecifiers: string[] = [];
+
+  if (parsed.program.body.length > 0) {
+    for (const stmt of parsed.program.body) {
+      if (stmt.type === "ImportDeclaration") {
+        const ref: ImportRef = {
+          specifier: stmt.source.value,
+          hasDefault: false,
+          hasNamespace: false,
+          named: [],
+        };
+        for (const spec of stmt.specifiers ?? []) {
+          if (spec.type === "ImportDefaultSpecifier") ref.hasDefault = true;
+          else if (spec.type === "ImportNamespaceSpecifier") ref.hasNamespace = true;
+          else ref.named.push(content.slice(spec.imported.start, spec.imported.end));
+        }
+        importRefs.push(ref);
+      } else if (stmt.type === "ExportNamedDeclaration" && stmt.source) {
+        reexportSpecifiers.push(stmt.source.value);
+      } else if (stmt.type === "ExportAllDeclaration") {
+        reexportSpecifiers.push(stmt.source.value);
+      }
+    }
+  } else {
+    // Fatal syntax error: the AST is empty but oxc's module record still
+    // carries the import/export shape — use it so import edges survive for
+    // files that are temporarily unparseable.
+    for (const staticImport of parsed.module.staticImports) {
+      const ref: ImportRef = {
+        specifier: staticImport.moduleRequest.value,
+        hasDefault: false,
+        hasNamespace: false,
+        named: [],
+      };
+      for (const entry of staticImport.entries) {
+        if (entry.importName.kind === "NamespaceObject") ref.hasNamespace = true;
+        else if (entry.importName.kind === "Default") ref.hasDefault = true;
+        else if (entry.importName.kind === "Name") {
+          ref.named.push(
+            entry.importName.start != null
+              ? content.slice(entry.importName.start, entry.importName.end ?? undefined)
+              : (entry.importName.name ?? ""),
+          );
+        }
+      }
+      importRefs.push(ref);
+    }
+    for (const staticExport of parsed.module.staticExports) {
+      const request = staticExport.entries.find((e) => e.moduleRequest != null)?.moduleRequest;
+      if (request) reexportSpecifiers.push(request.value);
+    }
+  }
+
   const sourceId = `file:${relPath}`;
   const useSymbol = mode === "symbol" && !isTest;
 
-  for (const decl of sourceFile.getImportDeclarations()) {
-    const moduleSpecifier = decl.getModuleSpecifierValue();
-    if (!moduleSpecifier.startsWith(".")) continue;
-
-    const resolved = resolveImport(sourceFile, decl);
+  // All import edges first, then re-export edges (two separate statement
+  // loops — matching the original backend's edge order).
+  for (const ref of importRefs) {
+    if (!ref.specifier.startsWith(".")) continue;
+    const resolved = resolveRelativeImport(filePath, ref.specifier, ctx);
     if (!resolved) continue;
-
     const targetRel = relative(rootDir, resolved);
 
     if (useSymbol) {
-      const namedImports = decl.getNamedImports();
-      const defaultImport = decl.getDefaultImport();
-      const namespaceImport = decl.getNamespaceImport();
-
-      if (namespaceImport) {
+      if (ref.hasNamespace) {
         edges.push({
           source: sourceId,
           target: `file:${targetRel}`,
@@ -244,7 +746,7 @@ function extractImports(
           provenances: ["ts-import"],
         });
       } else {
-        if (defaultImport) {
+        if (ref.hasDefault) {
           edges.push({
             source: sourceId,
             target: `symbol:${targetRel}#default`,
@@ -252,8 +754,7 @@ function extractImports(
             provenances: ["ts-import"],
           });
         }
-        for (const named of namedImports) {
-          const importName = named.getAliasNode() ? named.getNameNode().getText() : named.getName();
+        for (const importName of ref.named) {
           edges.push({
             source: sourceId,
             target: `symbol:${targetRel}#${importName}`,
@@ -261,7 +762,7 @@ function extractImports(
             provenances: ["ts-import"],
           });
         }
-        if (!defaultImport && namedImports.length === 0 && !namespaceImport) {
+        if (!ref.hasDefault && ref.named.length === 0) {
           edges.push({
             source: sourceId,
             target: `file:${targetRel}`,
@@ -280,31 +781,293 @@ function extractImports(
     }
   }
 
-  for (const exportDecl of sourceFile.getExportDeclarations()) {
-    const moduleSpecifier = exportDecl.getModuleSpecifierValue();
-    if (moduleSpecifier?.startsWith(".")) {
-      const resolved = resolveImport(sourceFile, exportDecl);
-      if (resolved) {
-        const targetRel = relative(rootDir, resolved);
-        edges.push({
-          source: sourceId,
-          target: `file:${targetRel}`,
-          kind: "imports",
-          provenances: ["ts-import"],
-        });
+  for (const specifier of reexportSpecifiers) {
+    if (!specifier.startsWith(".")) continue;
+    const resolved = resolveRelativeImport(filePath, specifier, ctx);
+    if (!resolved) continue;
+    const targetRel = relative(rootDir, resolved);
+    edges.push({
+      source: sourceId,
+      target: `file:${targetRel}`,
+      kind: "imports",
+      provenances: ["ts-import"],
+    });
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Relative module specifier resolution
+// ---------------------------------------------------------------------------
+
+// Differential probing against the TS checker (via ts-morph) across tsconfig
+// variants showed the resolution outcome depends only on these three options.
+// In particular moduleResolution (node16 vs node10 vs bundler) and
+// package.json "type" do NOT change it — the checker resolves extensionless
+// relative specifiers even under node16, and .tsx / .js / .json availability
+// is gated purely by jsx / allowJs / resolveJsonModule.
+interface ResolverContext {
+  jsx: boolean;
+  allowJs: boolean;
+  resolveJsonModule: boolean;
+  moduleCheckCache: Map<string, boolean>;
+}
+
+function createResolverContext(rootDir: string): ResolverContext {
+  // Only <rootDir>/tsconfig.json is consulted (no upward walk); without it,
+  // compiler defaults apply (all three off).
+  const options = readTsconfigResolveOptions(resolve(rootDir, "tsconfig.json"));
+  return { ...options, moduleCheckCache: new Map() };
+}
+
+function resolveRelativeImport(
+  fromFile: string,
+  specifier: string,
+  ctx: ResolverContext,
+): string | undefined {
+  const base = resolve(dirname(fromFile), specifier);
+  // `./dir/` is directory-only; `./x` tries file forms first, then directory
+  // (a directory literally named `x.js` still resolves via its index).
+  let resolved = specifier.endsWith("/") ? undefined : resolveAsFile(base, ctx);
+  if (resolved === undefined && isDirectory(base)) {
+    resolved = resolveAsDirectory(base, ctx);
+  }
+  if (resolved === undefined) return undefined;
+  // The checker only hands out a symbol for MODULE targets: a script target
+  // (no import/export syntax) produces no edge. JSON modules are implicit.
+  if (!resolved.endsWith(".json") && !isModuleFile(resolved, ctx)) return undefined;
+  return resolved;
+}
+
+// Extension probe order replicated from TS resolution behavior:
+// .ts, .tsx, .d.ts, .js (allowJs), .jsx (jsx AND allowJs). `.tsx` is ALWAYS
+// probed, but without the jsx option a .tsx match is discarded — and shadows
+// every lower-priority candidate (an x.tsx on disk hides x.d.ts and x.js).
+function probeSubstitution(base: string, ctx: ResolverContext): string | undefined {
+  const extensions = [".ts", ".tsx", ".d.ts"];
+  if (ctx.allowJs) extensions.push(".js");
+  if (ctx.allowJs && ctx.jsx) extensions.push(".jsx");
+  for (const extension of extensions) {
+    if (isFile(base + extension)) {
+      if (extension === ".tsx" && !ctx.jsx) return undefined;
+      return base + extension;
+    }
+  }
+  return undefined;
+}
+
+function probeExtensions(base: string, extensions: string[]): string | undefined {
+  for (const extension of extensions) {
+    if (isFile(base + extension)) return base + extension;
+  }
+  return undefined;
+}
+
+function resolveAsFile(candidate: string, ctx: ResolverContext): string | undefined {
+  if (candidate.endsWith(".tsx")) {
+    return ctx.jsx && isFile(candidate) ? candidate : undefined;
+  }
+  if (/\.[mc]?ts$/.test(candidate)) {
+    // .ts / .mts / .cts (and .d.ts / .d.mts / .d.cts): exact file only.
+    return isFile(candidate) ? candidate : undefined;
+  }
+  if (candidate.endsWith(".json")) {
+    return ctx.resolveJsonModule && isFile(candidate) ? candidate : undefined;
+  }
+  if (/\.jsx?$/.test(candidate)) {
+    return probeSubstitution(candidate.replace(/\.jsx?$/, ""), ctx);
+  }
+  if (candidate.endsWith(".mjs")) {
+    const extensions = [".mts", ".d.mts"];
+    if (ctx.allowJs) extensions.push(".mjs");
+    return probeExtensions(candidate.slice(0, -".mjs".length), extensions);
+  }
+  if (candidate.endsWith(".cjs")) {
+    const extensions = [".cts", ".d.cts"];
+    if (ctx.allowJs) extensions.push(".cjs");
+    return probeExtensions(candidate.slice(0, -".cjs".length), extensions);
+  }
+  // Extensionless (or an unrecognized extension): append and probe.
+  // .mts/.cts are deliberately NOT probed here — tsc doesn't either.
+  return probeSubstitution(candidate, ctx);
+}
+
+function resolveAsDirectory(dir: string, ctx: ResolverContext): string | undefined {
+  const pkgPath = resolve(dir, "package.json");
+  if (isFile(pkgPath)) {
+    let pkg: unknown;
+    try {
+      pkg = JSON.parse(readFileSync(pkgPath, "utf-8"));
+    } catch {
+      pkg = undefined;
+    }
+    if (pkg && typeof pkg === "object") {
+      const fields = pkg as { types?: unknown; typings?: unknown; main?: unknown };
+      const types =
+        typeof fields.types === "string"
+          ? fields.types
+          : typeof fields.typings === "string"
+            ? fields.typings
+            : undefined;
+      // A present-but-unresolvable "types" field suppresses "main" and falls
+      // through to index resolution (tsc consults main only when no types
+      // field exists at all).
+      const target = types ?? (typeof fields.main === "string" ? fields.main : undefined);
+      if (target !== undefined) {
+        const resolved = resolveAsFile(resolve(dir, target), ctx);
+        if (resolved) return resolved;
       }
+    }
+  }
+  return probeSubstitution(resolve(dir, "index"), ctx);
+}
+
+// A resolved target only yields an import edge when TS treats it as a module:
+// it has ESM syntax (import/export/import.meta — tracked by oxc's
+// error-tolerant module record even for broken files), `export =`, or
+// `import x = require(...)`.
+function isModuleFile(filePath: string, ctx: ResolverContext): boolean {
+  const cached = ctx.moduleCheckCache.get(filePath);
+  if (cached !== undefined) return cached;
+
+  let result = false;
+  try {
+    const content = stripBom(readFileSync(filePath, "utf-8"));
+    const parsed = loadOxc().parseSync(filePath, content);
+    result =
+      parsed.module.hasModuleSyntax ||
+      parsed.program.body.some(
+        (stmt) =>
+          stmt.type === "TSExportAssignment" ||
+          (stmt.type === "TSImportEqualsDeclaration" &&
+            stmt.moduleReference.type === "TSExternalModuleReference"),
+      );
+  } catch {
+    result = false;
+  }
+  ctx.moduleCheckCache.set(filePath, result);
+  return result;
+}
+
+// ---------------------------------------------------------------------------
+// tsconfig reading (jsx / allowJs / resolveJsonModule only)
+// ---------------------------------------------------------------------------
+
+const JSX_VALUES = new Set(["preserve", "react", "react-native", "react-jsx", "react-jsxdev"]);
+
+interface ResolveOptions {
+  jsx: boolean;
+  allowJs: boolean;
+  resolveJsonModule: boolean;
+}
+
+function readTsconfigResolveOptions(tsconfigPath: string): ResolveOptions {
+  // Collect the extends chain leaf-first, then apply base-to-leaf so nearer
+  // configs override. Known limit: "extends" is resolved best-effort
+  // (relative paths and a node_modules walk for package-style values);
+  // exotic setups (extends arrays, "rootDirs", package "exports"-mapped
+  // configs) are not followed.
+  const chain: Array<Record<string, unknown>> = [];
+  const visited = new Set<string>();
+  let current: string | undefined = tsconfigPath;
+  while (current !== undefined && !visited.has(current) && isFile(current)) {
+    visited.add(current);
+    const parsed = parseJsonc(readFileSync(current, "utf-8"));
+    if (!parsed || typeof parsed !== "object") break;
+    const config = parsed as Record<string, unknown>;
+    chain.push(config);
+    current =
+      typeof config.extends === "string"
+        ? resolveExtendsPath(dirname(current), config.extends)
+        : undefined;
+  }
+
+  const options: ResolveOptions = { jsx: false, allowJs: false, resolveJsonModule: false };
+  for (let i = chain.length - 1; i >= 0; i--) {
+    const compilerOptions = chain[i].compilerOptions;
+    if (!compilerOptions || typeof compilerOptions !== "object") continue;
+    const co = compilerOptions as Record<string, unknown>;
+    if (co.jsx !== undefined) {
+      options.jsx = typeof co.jsx === "string" && JSX_VALUES.has(co.jsx.toLowerCase());
+    }
+    if (co.allowJs !== undefined) options.allowJs = co.allowJs === true;
+    if (co.resolveJsonModule !== undefined) {
+      options.resolveJsonModule = co.resolveJsonModule === true;
+    }
+  }
+  return options;
+}
+
+function resolveExtendsPath(fromDir: string, extendsValue: string): string | undefined {
+  const tryJson = (p: string): string | undefined =>
+    isFile(p) ? p : !p.endsWith(".json") && isFile(`${p}.json`) ? `${p}.json` : undefined;
+  if (extendsValue.startsWith(".") || extendsValue.startsWith("/")) {
+    return tryJson(resolve(fromDir, extendsValue));
+  }
+  // Package-style extends ("@tsconfig/node22"): best-effort node_modules
+  // lookup from the config's directory upward.
+  let dir = fromDir;
+  for (;;) {
+    const base = resolve(dir, "node_modules", extendsValue);
+    const found = tryJson(base) ?? tryJson(resolve(base, "tsconfig.json"));
+    if (found) return found;
+    const parent = dirname(dir);
+    if (parent === dir) return undefined;
+    dir = parent;
+  }
+}
+
+// tsconfig.json is JSONC: strip comments (string-aware), then parse; retry
+// with trailing commas removed if strict parsing fails.
+function parseJsonc(text: string): unknown {
+  let out = "";
+  let i = 0;
+  while (i < text.length) {
+    const ch = text[i];
+    if (ch === '"') {
+      out += ch;
+      i++;
+      while (i < text.length) {
+        out += text[i];
+        if (text[i] === "\\") {
+          out += text[i + 1] ?? "";
+          i += 2;
+          continue;
+        }
+        if (text[i] === '"') {
+          i++;
+          break;
+        }
+        i++;
+      }
+      continue;
+    }
+    if (ch === "/" && text[i + 1] === "/") {
+      while (i < text.length && text[i] !== "\n") i++;
+      continue;
+    }
+    if (ch === "/" && text[i + 1] === "*") {
+      i += 2;
+      while (i < text.length && !(text[i] === "*" && text[i + 1] === "/")) i++;
+      i += 2;
+      continue;
+    }
+    out += ch;
+    i++;
+  }
+  try {
+    return JSON.parse(out);
+  } catch {
+    try {
+      return JSON.parse(out.replace(/,(\s*[}\]])/g, "$1"));
+    } catch {
+      return undefined;
     }
   }
 }
 
-function resolveImport(sourceFile: SourceFile, decl: any): string | undefined {
-  try {
-    const resolved = decl.getModuleSpecifierSourceFile?.();
-    return resolved?.getFilePath();
-  } catch {
-    return undefined;
-  }
-}
+// ---------------------------------------------------------------------------
+// Requirement-ID tag extraction (plain-text regex — needs no AST)
+// ---------------------------------------------------------------------------
 
 export function extractImplTags(
   content: string,
@@ -387,6 +1150,45 @@ function resolveSymbolAtLine(ranges: SymbolRange[], line: number): string | null
     }
   }
   return best?.name ?? null;
+}
+
+// ---------------------------------------------------------------------------
+// Line numbers (1-based, counted at the trivia-free span offsets) and misc
+// ---------------------------------------------------------------------------
+
+function buildLineStarts(content: string): number[] {
+  const starts = [0];
+  for (let i = 0; i < content.length; i++) {
+    if (content[i] === "\n") starts.push(i + 1);
+  }
+  return starts;
+}
+
+function lineOf(lineStarts: number[], offset: number): number {
+  let lo = 0;
+  let hi = lineStarts.length - 1;
+  while (lo < hi) {
+    const mid = (lo + hi + 1) >> 1;
+    if (lineStarts[mid] <= offset) lo = mid;
+    else hi = mid - 1;
+  }
+  return lo + 1;
+}
+
+function isFile(p: string): boolean {
+  try {
+    return statSync(p).isFile();
+  } catch {
+    return false;
+  }
+}
+
+function isDirectory(p: string): boolean {
+  try {
+    return statSync(p).isDirectory();
+  } catch {
+    return false;
+  }
 }
 
 export function hash(content: string): string {
